microservice/email_service/requirements.txt
pika==1.3.2
python-dotenv==1.0.1
pytest==8.2.1
